Change float to decimal('price',8,2)

+++ database/migrations/2020_01_20_143018_create_dishes_table.php
@@ -17,7 +17,7 @@ class CreateDishesTable extends Migration
             $table->bigIncrements('id');
         //I need to add this    $table->string('slug');
             $table->string('name');
-            $table->float('price')->nullable();
+            $table->decimal('price', 8, 2)->nullable();
             $table->text('description')->nullable();
             $table->string('image')->nullable();
             $table->timestamps();
